+++ packages/lumx-vue/src/components/input-required-legend/index.ts
@@ -1,0 +1,4 @@
+import InputRequiredLegend, { InputRequiredLegendProps } from './InputRequiredLegend';
+
+export { InputRequiredLegend };
+export type { InputRequiredLegendProps };