+++ packages/site-demo/content/product/components/input-required-legend/vue/default.vue
@@ -1,0 +1,9 @@
+<template>
+    <InputRequiredLegend label="Indicates a required field" :theme="theme" />
+</template>
+
+<script setup lang="ts">
+import { InputRequiredLegend, type Theme } from '@lumx/vue';
+
+defineProps<{ theme?: Theme }>();
+</script>